requirements.txt
annotated-types==0.7.0
anyio==4.7.0
audioop-lts==0.2.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
MarkupSafe==3.0.2
moviepy==2.1.1
nltk==3.9.1
numpy==2.2.0
openai==1.58.1
pillow==10.4.0
proglog==0.1.10
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
python-telegram-bot==21.9
regex==2024.11.6
requests==2.32.3
setuptools==75.6.0
sniffio==1.3.1
SpeechRecognition==3.12.0
SQLAlchemy==2.0.36
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
